Fix format specifier warning at src/pet/petmem.c:448:25
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36546 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/pet/petmem.c
@@ -445,7 +445,7 @@ static int efe0_dump(void)
     int i;
     int mask = 1;
 
-    mon_out("efe0 = $%02x; previous in = $%02x; odd/even = %d\n", dongle6702.val, dongle6702.prevodd, dongle6702.wantodd);
+    mon_out("efe0 = $%02x; previous in = $%02x; odd/even = %d\n", (unsigned int)dongle6702.val, dongle6702.prevodd, dongle6702.wantodd);
     for (i = 0; i < 8; i++, mask <<= 1) {
         int j;
         int maskj;
